+++ CHANGELOG.md
@@ -11,12 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `asWrapper` function to cast native refs to wrapper refs (#30, #31)
 - `DataValue::getValuePtr` method
+- `Server::getNamespaceArray` method
 
 ### Changed
 
 - Templated `Node` class (`Node<Server>`, `Node<Client>`) as preparation for client implementation (#32)
 - Return reference from `ExpandedNodeId::getNodeId`
 - Pass `LocalizedText` instead of members (`locale`, `text`) to `Node::writeDisplayName`, `Node::writeDescription`, `services::writeDisplayName`, `services::writeDescription` (#29)
+- Remove `Server::getConfig` method
 
 ### Fixed